requirements.txt
alembic==1.13.2
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
eralchemy==1.5.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
graphviz==0.20.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
msgspec==0.18.6
packaging==24.1
pluggy==1.5.0
probableparsing==0.0.1
pytest==8.3.3
python-crfsuite==0.9.11
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
usaddress==0.5.11
Werkzeug==3.0.4
